% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_entropy.R
\name{cross_entropy}
\alias{cross_entropy}
\title{Cross Entropy Loss}
\usage{
cross_entropy(x, z)
}
\arguments{
\item{x}{The input data.}

\item{z}{The predicted output data.}
}
\value{
Cross entropy value.
}
\description{
Calculates the cross entropy loss between the input and output layers.
}
