% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deep_learning_model.R
\name{deep_learning_model}
\alias{deep_learning_model}
\title{Deep Learning Model}
\usage{
deep_learning_model(x, W, b)
}
\arguments{
\item{x}{A matrix of input features.}

\item{W}{A matrix of weights.}

\item{b}{A bias vector.}
}
\value{
The activation values of the hidden layers.
}
\description{
Implements a deep learning model with sigmoid activation and multiple hidden
layers.
}
