% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_models.R
\name{evaluate_models}
\alias{evaluate_models}
\title{Evaluate Models}
\usage{
evaluate_models(models, data, target)
}
\arguments{
\item{models}{List of trained models.}

\item{data}{The test data.}

\item{target}{The target variable in the test data.}
}
\value{
AUC values for each model.
}
\description{
Evaluates the performance of models using ROC and AUC.
}
