% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_importance.R
\name{feature_importance}
\alias{feature_importance}
\title{Feature Importance}
\usage{
feature_importance(model)
}
\arguments{
\item{model}{A trained model.}
}
\value{
A dataframe of feature importance.
}
\description{
Ranks the features based on their importance in the trained models.
}
