% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_dl_model.R
\name{train_dl_model}
\alias{train_dl_model}
\title{Train Deep Learning Model}
\usage{
train_dl_model(
  data,
  target,
  epochs = 100,
  l1 = 0.001,
  l2 = 0.001,
  input_dropout_ratio = 0.2
)
}
\arguments{
\item{data}{A dataframe or matrix of input features.}

\item{target}{The target variable for prediction.}

\item{epochs}{The number of training epochs.}

\item{l1}{The L1 regularization parameter.}

\item{l2}{The L2 regularization parameter.}

\item{input_dropout_ratio}{The dropout ratio for the input layer.}
}
\value{
Trained h2o model.
}
\description{
Trains the deep learning model using the h2o package.
}
