% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_ml_model.R
\name{train_ml_models}
\alias{train_ml_models}
\title{Train Machine Learning Models}
\usage{
train_ml_models(data, target)
}
\arguments{
\item{data}{A dataframe or matrix of input features.}

\item{target}{The target variable for prediction.}
}
\value{
List of trained models.
}
\description{
Trains multiple machine learning models using the caret package.
}
